==> main/01_fdm_file_management.ipynb <==
# FDM File Management
1. Convert CSV dataset to Parquet to reduce file size. - deprecated
2. Create database of list of FDM files, origin file, path, filesize info. 

## Import

import pandas as pd
import numpy as np
import os
import math

### Constant

fdm_folder_path = '../dataset/fdm'
output = '../dataset/database_local/01_fdm_files.csv'

# fdm_folder_path = '/Volumes/My Passport/desy'
# output = '../dataset/database/01_fdm_files.csv'

## Read Files
Retrieve all FDM files information and create dataset.

# Scan through the folder
filename_list, tail_id_list, fsize_mb_list, fullpath_list = [], [], [], []
for root, dirs, files in os.walk(fdm_folder_path):
    for file in files:
        if file.endswith(".csv"):
            path = os.path.join(root, file)
            filesize = round(os.stat(path).st_size / 1000000, 2)
            tail_id = os.path.basename(root)

            filename_list.append(file)
            fullpath_list.append(path)
            fsize_mb_list.append(filesize)
            tail_id_list.append(tail_id)

### Create File Database

col_to_list = {'fname': filename_list,
               'tail_id': tail_id_list,
               'fsize_mb': fsize_mb_list,
               'fullpath': fullpath_list}
fdm_files = pd.DataFrame(col_to_list)

fdm_files

fdm_files.fname.is_unique

fdm_files.to_csv(output, index=False)
