==> main/02_fdm_departure_arrival_location.ipynb <==
# FDM Departure and Arrival
Finding departure and arrival coordinate for each FDM files. Then, create `fdm_location` dataset using the information.

## Import

import pandas as pd
import seaborn as sns
import numpy as np
import gc
import os

### Constant

# fdm_database_path = '../dataset/database/01_fdm_files.csv'
# fdm_locations_temp = "../dataset/database/fdm_location_write.csv"
# fdm_locations_output = '../dataset/database/02_fdm_locations.csv'

fdm_database_path = '../dataset/database_local/01_fdm_files.csv'
fdm_locations_temp = "../dataset/database_local/fdm_location_write.csv"
fdm_locations_output = '../dataset/database_local/02_fdm_locations.csv'

## Read Data

#### 1. Read FDM Management Data

fdm_files = pd.read_csv(fdm_database_path)
display(fdm_files.head())
print(fdm_files.shape)

## Find Departure Coordinate

#### 1. Create Helper Function

def get_first_loc(fdm_df):
    # Given FDM Dataframe, get location (lat, lon) from first row
    if fdm_df.empty:
        return (np.nan, np.nan)
    else:
        return tuple(fdm_df[['lat_rad', 'lon_rad']].iloc[0])

def get_last_loc(fdm_df):
    # Given FDM Dataframe, get location (lat, lon) from last row
    if fdm_df.empty:
        return (np.nan, np.nan)
    else:
        return tuple(fdm_df[['lat_rad', 'lon_rad']].iloc[-1])

def get_departure_loc(fdm):
    # Given fdm dataframe, return the departure long lat and arrival long lat
    plane_in_ground   = fdm[fdm['wow'] == 0]
    plane_off_ground  = fdm[fdm['wow'] == 1]
        
    if plane_in_ground.empty:
        return (np.nan, np.nan)

    if plane_off_ground.empty:
        est_departure_loc = tuple(plane_in_ground[['lat_rad', 'lon_rad']].iloc[0])
        return est_departure_loc
    else:
        first_takeoff_idx = plane_off_ground.index[0]
        departure_df = plane_in_ground.loc[:first_takeoff_idx]
        est_departure_loc = get_first_loc(departure_df)
        return est_departure_loc

def get_arrival_loc(fdm):
    # Given fdm dataframe, return the departure long lat and arrival long lat
    plane_in_ground   = fdm[fdm['wow'] == 0]
    plane_off_ground  = fdm[fdm['wow'] == 1]
        
    if plane_in_ground.empty:
        return (np.nan, np.nan)

    if plane_off_ground.empty:
        return (np.nan, np.nan)
    else:
        first_takeoff_idx = plane_off_ground.index[0]
        arrival_df = plane_in_ground.loc[first_takeoff_idx:]
        est_arrival_loc = get_last_loc(arrival_df)
        return est_arrival_loc

locs = []
def get_loc_from_fdm(fdm_filepath, write=True):
    # Given filepath, return the departure long lat and arrival long lat    
    fdm_df = pd.read_csv(fdm_filepath)
    fdm_df = fdm_df[(fdm_df['lat_rad'] != 0) & (fdm_df['lon_rad'] != 0)]
    dep_lat, dep_lon, arr_lat, arr_lon = get_departure_loc(fdm_df) + get_arrival_loc(fdm_df)
    
    row = (os.path.basename(fdm_filepath), dep_lat, dep_lon, arr_lat, arr_lon)
    locs.append(row)
    
    if write:
        file1 = open(fdm_locations_temp, "a")  # append mode
        file1.write(",".join(str(r) for r in row)  + '\n')
        file1.close()
    
    print(os.path.basename(fdm_filepath), end=': ')
    print((round(dep_lat, 4), round(dep_lon, 4), round(arr_lat, 4), round(arr_lon, 4)), end=' | ')
    print(len(locs), '/', len(fdm_files))
    return (dep_lat, dep_lon, arr_lat, arr_lon)

fpath = fdm_files.fullpath[0]
get_loc_from_fdm(fpath, write=False)

#### 2. Generate Location for All FDM Files

loc_columns = ['dep_lat', 'dep_lon', 'arr_lat', 'arr_lon']
csv_header = ['fname'] + loc_columns

file1 = open(fdm_locations_temp, "w")
L = ','.join(csv_header) + '\n'
file1.writelines(L)
file1.close()

locs = []
loc_tuples = fdm_files[:5].fullpath.map(get_loc_from_fdm)

fdm_locations = fdm_files[['fname']].copy()
fdm_locations[loc_columns] = pd.DataFrame(loc_tuples.tolist(), index=fdm_files.index)

fdm_locations.head()

fdm_locations.to_csv(fdm_locations_output, index=False)
