==> main/03_airport_detection.ipynb <==
# Airport Detection
Detect departure and arrival airport from Flight Data Monitoring dataset.
Using KNN model to match longitude and latitude of airport.

## Import

%load_ext autoreload
%autoreload 2

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import math

from sklearn.neighbors import KNeighborsClassifier

### Constant

# fdm_database_path = '../dataset/database/01_fdm_files.csv'
# fdm_locations_path = '../dataset/database/02_fdm_location.csv'
# flights_airport_output = '../dataset/database/03_flights_airport.csv'

fdm_database_path = '../dataset/database_local/01_fdm_files.csv'
fdm_locations_path = '../dataset/database_local/02_fdm_locations.csv'
flights_airport_output = '../dataset/database_local/03_flights_airport.csv'

## Read Data

#### 1. Read Airport Data

airports_filepath = '../dataset/database/airports (radians).xlsx'
airports = pd.read_excel(airports_filepath, index_col=0)
airports.head()

#### 2. Read FDM File Data

fdm_files = pd.read_csv(fdm_database_path)
fdm_files.head()

#### 3. Read FDM Location

fdm_locations = pd.read_csv(fdm_locations_path)
fdm_locations.head()

fdm_locations_cleaned = fdm_locations[~fdm_locations.isnull().any(axis=1)]
fdm_locations_cleaned.shape

## Exploratory Data Analysis

#### 1. Create Airport Locations

airport_locations = airports[['iata_code', 'latitude_rad', 'longitude_rad']]
airport_locations.head()

airport_locations = airport_locations.dropna()

airport_locations.info()

No null value detected

## Model Training

knn = KNeighborsClassifier(n_neighbors=1)

X_train = airport_locations[['latitude_rad', 'longitude_rad']]
y_train = airport_locations['iata_code']

knn.fit(X_train, y_train)

### Predict Airport

#### 1. Departure Airport

departure_list = knn.predict(fdm_locations_cleaned[['dep_lat', 'dep_lon']])
departure_list

#### 2. Arrival Airport

arrival_list = knn.predict(fdm_locations_cleaned[['arr_lat', 'arr_lon']])
arrival_list

#### Join Prediction to Table

fdm_airport = fdm_locations_cleaned[['fname']].copy()
fdm_airport['dep_airport'] = departure_list
fdm_airport['arr_airport'] = arrival_list

fdm_airport.head()

#### Merge with FDM Files

fdm_files

fdm_files_merged = fdm_files.copy()
fdm_files_merged = fdm_files_merged.join(fdm_airport.set_index('fname'),
                                         on='fname',
                                         how='left')

fdm_files_merged.head()

#### Reorder column

col_order = ['fname', 'tail_id', 'dep_airport', 'arr_airport', 'fsize_mb', 'fullpath']
fdm_files_merged = fdm_files_merged[col_order]

fdm_files_merged.head()

fdm_files_merged.to_csv(flights_airport_output, index=False)
