==> main/04_runway_detection.ipynb <==
# Runway Detection
Detect departure and arrival runway from Flight Data Monitoring dataset. Using KNN model to match longitude and latitude of the runway.

## Import 

%load_ext autoreload
%autoreload 2

import pandas as pd
import math
import numpy as np
import seaborn as sns
import matplotlib as plt
import folium
from sklearn.neighbors import KNeighborsClassifier

### Constant

base_path = '../dataset/database_local'
# base_path = '../dataset/database'

fdm_database_path = base_path + '/01_fdm_files.csv'
fdm_locations_path = base_path + '/02_fdm_locations.csv'
flights_path = base_path + '/03_flights_airport.csv'
landing_output = base_path + '/04_landing_count_detail.csv'
flights_runway_output = base_path + '/04_flights_with_runway.csv'

## Extract Data

#### 1. Read Flights Data 

flights = pd.read_csv(flights_path)
display(flights.head())
flights.shape

flights_clean = flights.dropna().copy()
flights_clean.shape

#### 2. Read Flight Location

fdm_locations = pd.read_csv(fdm_locations_path)
fdm_locations.head()

fdm_locations_clean = fdm_locations.dropna().copy()
fdm_locations_clean.shape

#### 3. Read Runway Data

runways_path = '../dataset/database/runways (radians).xlsx' 
runways = pd.read_excel(runways_path, index_col=0)
runways.head()

#### 4. Read Airport Data

airports_filepath = '../dataset/database/airports (radians).xlsx'
airports = pd.read_excel(airports_filepath, index_col=0)
airports.head()

## Transform Data

#### 1. Concat Runway Identifier

runways_le_ident = runways[['iata_code', 'le_ident', 'le_latitude_rad', 'le_longitude_rad']].copy()
runways_he_ident = runways[['iata_code', 'he_ident', 'he_latitude_rad', 'he_longitude_rad']].copy()

# Rename column to be the same
le_col_rename = {'le_ident': 'runway_ident', 'le_latitude_rad': 'lat_rad', 'le_longitude_rad': 'lon_rad'}
he_col_rename = {'he_ident': 'runway_ident', 'he_latitude_rad': 'lat_rad', 'he_longitude_rad': 'lon_rad'}

runways_le_ident = runways_le_ident.rename(columns=le_col_rename)
runways_he_ident = runways_he_ident.rename(columns=he_col_rename)

runways_le_ident['type'] = 'le'
runways_he_ident['type'] = 'he'

display(runways_le_ident.head())
display(runways_he_ident.head())

# Concat
runways_frames = [runways_le_ident, runways_he_ident]
runways_ident = pd.concat(runways_frames, ignore_index=False)
runways_ident = runways_ident.sort_index().reset_index().rename(columns={'no': 'rwy_no'})
runways_ident.head()

#### 2. Create New Runway ID with IATA Code

rwy_aprt_id = runways_ident['iata_code'] + '.' + runways_ident['runway_ident']
runways_ident['rwy_aprt_id'] = rwy_aprt_id

runways_ident.head(5)

## Train Model

cleaned_runways = runways_ident.dropna().copy()

knn = KNeighborsClassifier(n_neighbors=1)

X_train = cleaned_runways[['lat_rad', 'lon_rad']]
y_train = cleaned_runways['rwy_aprt_id']

knn.fit(X_train, y_train)

## Predict Runway

departure_list = knn.predict(fdm_locations_clean[['dep_lat', 'dep_lon']])
departure_list

arrival_list = knn.predict(fdm_locations_clean[['arr_lat', 'arr_lon']])
arrival_list

#### Merge with FDM filename column

flights_runway = fdm_locations_clean[['fname']].copy()
flights_runway['dep_runway'] = departure_list
flights_runway['arr_runway'] = arrival_list
flights_runway[['dep_aprt', 'dep_rwy']] = flights_runway['dep_runway'].str.split('.',expand=True)
flights_runway[['arr_aprt', 'arr_rwy']] = flights_runway['arr_runway'].str.split('.',expand=True)
flights_runway.drop(columns=['dep_runway', 'arr_runway'], inplace=True)

flights_runway.head()

## Check Airport Result
Check if airport result from module 03 is the same with the result

flights_airport = flights_clean[['fname', 'dep_airport', 'arr_airport']]
flights_joined = flights_airport.join(flights_runway.set_index('fname'), on='fname', how='outer')
flights_joined.head()

len(flights_joined[flights_joined.any(1).isnull()])

sum(flights_joined['dep_airport'] != flights_joined['dep_aprt'])

flights_joined[flights_joined['dep_airport'] != flights_joined['dep_aprt']]

sum(flights_joined['arr_airport'] != flights_joined['arr_aprt'])

flights_joined[flights_joined['arr_airport'] != flights_joined['arr_aprt']]

# Set runway to null if airport result is diffeent
dep_rwy = flights_joined.apply(lambda df: df['dep_rwy'] if df['dep_airport'] == df['dep_aprt'] else None, axis=1)
arr_rwy = flights_joined.apply(lambda df: df['arr_rwy'] if df['arr_airport'] == df['arr_aprt'] else None, axis=1)
flights_joined['dep_rwy'] = dep_rwy
flights_joined['arr_rwy'] = arr_rwy

flights_joined[flights_joined.isna().any(1)]

flights_appended = flights.join(flights_joined[['fname', 'dep_rwy', 'arr_rwy']].set_index('fname'),
                                on='fname',
                                how='outer')
col_order = ['fname', 'tail_id', 'dep_airport', 'dep_rwy', 'arr_airport', 'arr_rwy', 'fsize_mb', 'fullpath']
flights_appended = flights_appended[col_order]

## Load to CSV

flights_appended.head()

flights_appended.to_csv(flights_runway_output, index=False)

## Find Most Runway Occurance

runways_merged = flights_appended.arr_airport + '.' + flights_appended.arr_rwy
runways_merged.value_counts()

most_landing = runways_merged.value_counts().index[0]
most_landing_aprt, most_landing_rwy = most_landing.split('.')

most_landing_aprt, most_landing_rwy

### Create DataFrame

# Create DataFrame
landing_count = pd.DataFrame(runways_merged.value_counts().reset_index())
landing_count.columns = ['airport_runway', 'flight_count']
landing_count.head()

airport_runway = landing_count["airport_runway"].str.split(".", n = 1, expand = True)
landing_count["iata_code"] = airport_runway[0]
landing_count["runway_ident"] = landing_count["airport_runway"]

landing_count.columns

col_order = ['iata_code', 'runway_ident', 'flight_count']
landing_count = landing_count[col_order]

landing_count.head()

### Merge with Airport Data 

landing_count_detail = landing_count.join(airports.set_index('iata_code'),
                                          on='iata_code',
                                          how='left')

landing_count_detail.head()

landing_count_detail.to_csv(landing_output, index=False)

## Visualize Runway

flight_msp_30r =flights_appended[(flights_appended['arr_airport'] == 'MSP') & 
                                 (flights_appended['arr_rwy'] == '30R')]

#### Get 5 Most Departure Airport to MSP.30R 

top_5_airport = flight_msp_30r['dep_airport'].value_counts()[:5].index
top_5_airport = list(top_5_airport)
print(top_5_airport)

#### Get Coordinate for each Airport 

coordinate_cols = ['iata_code', 'latitude_rad', 'longitude_rad']
top_5_coordinate = airports[airports['iata_code'].isin(top_5_airport)][coordinate_cols]

arrival_coordinate = runways[(runways.iata_code == 'MSP') & (runways.he_ident == '30R')]
arr_lat, arr_lon = arrival_coordinate[['he_latitude_rad', 'he_longitude_rad']].iloc[0].values

top_5_coordinate[['arr_lat', 'arr_lon']] = arr_lat, arr_lon
top_5_coordinate = top_5_coordinate.rename(columns={'latitude_rad': 'dep_lat', 'longitude_rad': 'dep_lon'})
top_5_coordinate = top_5_coordinate.set_index('iata_code', drop=True)
top_5_coordinate

top_5_records = top_5_coordinate.to_records()
top_5_records

#### Visualize Using Map

flight_msp_30r_loc = fdm_locations[fdm_locations.fname.isin(flight_msp_30r.fname)]
flight_msp_30r_loc = flight_msp_30r_loc.set_index('fname')
flight_msp_30r_loc[:5]

msp_coordinate = airports[airports['iata_code'] == 'MSP'][['latitude_rad', 'longitude_rad']].iloc[0].values
msp_coordinate = [math.degrees(coor) for coor in msp_coordinate]
print(msp_coordinate)

msp_runways_coordinate = runways_ident[runways_ident['iata_code'] == 'MSP']
msp_runways_coordinate

arr_coordinate = flight_msp_30r_loc[['arr_lat', 'arr_lon']].values
arr_coordinate = [(math.degrees(coor[0]), math.degrees(coor[1])) for coor in arr_coordinate]
arr_coordinate[:5]

m = folium.Map(location=msp_coordinate, zoom_start=14, tiles="Stamen Terrain")

for coor in arr_coordinate:
    folium.Circle(
        radius=1,
        location=coor,
        color="blue",
        fill=False,
    ).add_to(m)

for row in msp_runways_coordinate.itertuples():    
    folium.Marker(
        location=(math.degrees(row.lat_rad), math.degrees(row.lon_rad)),
        tooltip=f"Runway {row.rwy_aprt_id} ({row.type})",
        popup=f"Runway {row.rwy_aprt_id} ({row.type})",
        icon=folium.Icon(color="green"),
    ).add_to(m)

m
